+++ Screen.h
@@ -10,8 +10,8 @@
 class Screen
 {
 public:
-	const int XMAX;
-	const int YMAX;
+	const size_t XMAX;
+	const size_t YMAX;
 	Screen(int mx, int my) :XMAX(mx), YMAX(my), buffer(std::make_unique<QImage>(XMAX, YMAX, QImage::Format_RGB32))
 	{
 		buffer->fill(QColor(150, 150, 150));
